[docs] Extract some API section components
